requirements.txt
pydantic
python-dotenv
langchain
langchain-openai
langchain-groq
langchain-core
langchain-community
beautifulsoup4
requests
requests-html
wikipedia
langgraph
unstructured[all-docs]
pillow
lxml
chromadb
poppler-utils
langchain-groq
langchain_astradb
duckduckgo-search
yfinance
duckduckgo-search
streamlit